Move DispatchLoop to dispatch::loop
Move all of the DispatchLoop and its related structs and enums to their
own module.
